+++ packages/expo-router/src/react-navigation/core/BaseNavigationContainer.tsx
@@ -1,5 +1,4 @@
 'use client';
-import isEqual from 'fast-deep-equal';
 import * as React from 'react';
 import { use } from 'react';
 
@@ -10,13 +9,11 @@ import {
   areUrlObjectsEqual,
   getRouteInfoFromState,
 } from '../../global-state/getRouteInfoFromState';
-import {
-  GlobalRoutesWithRemovalPreventedContext,
-  RemovalPreventionProvider,
-} from '../../global-state/removalPrevention';
+import { GlobalRoutesWithRemovalPreventedContext } from '../../global-state/removalPrevention';
 import { RouteInfoContext } from '../../global-state/routeInfoContext';
 import { RouterConfigContext } from '../../global-state/routerConfigContext';
-import { RouterRegistryContext, RouterRegistryProvider } from '../../global-state/routerRegistry';
+import { RouterRegistryContext } from '../../global-state/routerRegistry';
+import { RoutingQueueApiContext } from '../../global-state/routingQueueContext';
 import { useNavigationTreeReducer } from '../../global-state/useNavigationTreeReducer';
 import { useNavigationTreeReportEvents } from '../../global-state/useNavigationTreeReportEvents';
 import useLatestCallback from '../../utils/useLatestCallback';
@@ -24,7 +21,6 @@ import {
   CommonActions,
   type InitialState,
   type NavigationAction,
-  type NavigationState,
   type ParamListBase,
   type Route,
 } from '../routers';
@@ -62,39 +58,19 @@ const duplicateNameWarnings: string[] = [];
  *
  * @param props.initialState Initial state object for the navigation tree.
  * @param props.onReady Callback which is called after the navigation tree mounts.
- * @param props.onStateChange Callback which is called with the latest navigation state when it changes.
  * @param props.onUnhandledAction Callback which is called when an action is not handled. TODO(@ubax): restore this callback. https://linear.app/expo/issue/ENG-26123
  * @param props.theme Theme object for the UI elements.
  * @param props.children Child elements to render the content.
  * @param props.ref Ref object which refers to the navigation object containing helper methods.
  */
 export function BaseNavigationContainer(props: InternalNavigationContainerProps) {
-  const registry = use(RouterRegistryContext);
-  const routesWithRemovalPrevented = use(GlobalRoutesWithRemovalPreventedContext);
-
-  // TODO(@ubax): investigate if this is really needed
-  let content = <BaseNavigationContainerInner {...props} />;
-  if (routesWithRemovalPrevented === undefined) {
-    content = <RemovalPreventionProvider>{content}</RemovalPreventionProvider>;
-  }
-  if (registry === undefined) {
-    content = <RouterRegistryProvider>{content}</RouterRegistryProvider>;
-  }
-  return content;
-}
-
-function BaseNavigationContainerInner({
-  ref,
-  initialState,
-  onStateChange,
-  onReady,
-  UNSTABLE_routeNode,
-  theme,
-  children,
-}: InternalNavigationContainerProps) {
+  const { ref, initialState, onReady, UNSTABLE_routeNode, theme, children } = props;
   const parent = use(NavigationStateContext);
   const inheritedRouteInfo = use(RouteInfoContext);
   const routerConfig = use(RouterConfigContext);
+  const routingQueue = use(RoutingQueueApiContext);
+  const registry = use(RouterRegistryContext);
+  const routesWithRemovalPrevented = use(GlobalRoutesWithRemovalPreventedContext);
 
   if (!parent.isDefault) {
     throw new Error(
@@ -102,8 +78,16 @@ function BaseNavigationContainerInner({
     );
   }
 
-  const registry = use(RouterRegistryContext)!;
-  const routesWithRemovalPrevented = use(GlobalRoutesWithRemovalPreventedContext)!;
+  if (
+    routingQueue === undefined ||
+    registry === undefined ||
+    routesWithRemovalPrevented === undefined
+  ) {
+    throw new Error(
+      'The navigation container requires the shared routing state provided by `ExpoRoot`. Render the navigation container inside `ExpoRoot`.'
+    );
+  }
+
   const emitter = useEventEmitter<NavigationContainerEventMap>();
 
   // TODO(@ubax): consider moving this state to ExpoRoot.
@@ -117,9 +101,6 @@ function BaseNavigationContainerInner({
       redirects: routerConfig?.redirects,
     });
   useNavigationTreeReportEvents(report, consumeReportEvents);
-
-  const hasNotifiedInitialStateRef = React.useRef(false);
-  const lastNotifiedStateRef = React.useRef<NavigationState | undefined>(undefined);
 
   const { listeners, addListener } = useChildListeners();
 
@@ -208,37 +189,13 @@ function BaseNavigationContainerInner({
 
   React.useImperativeHandle(ref, () => navigation, [navigation]);
 
-  const lastEmittedOptionsRef = React.useRef<
-    { options: object; routeKey: string | undefined } | undefined
-  >(undefined);
-
-  // TODO(@ubax): investigate if there is better way to implemnet this and wether this is really needed,
-  const onOptionsChange = useLatestCallback((options: object, routeKey?: string) => {
-    const lastEmittedOptions = lastEmittedOptionsRef.current;
-    if (
-      lastEmittedOptions?.routeKey === routeKey &&
-      lastEmittedOptions !== undefined &&
-      isEqual(lastEmittedOptions.options, options)
-    ) {
-      return;
-    }
-
-    lastEmittedOptionsRef.current = { options, routeKey };
-
-    emitter.emit({
-      type: 'options',
-      data: { options },
-    });
-  });
-
   const builderContext = React.useMemo(
     () => ({
       addListener,
       handleAction,
       resetNavigator,
-      onOptionsChange,
     }),
-    [addListener, handleAction, onOptionsChange, resetNavigator]
+    [addListener, handleAction, resetNavigator]
   );
 
   const context = React.useMemo(
@@ -260,10 +217,8 @@ function BaseNavigationContainerInner({
   }
 
   const onReadyRef = React.useRef(onReady);
-  const onStateChangeRef = React.useRef(onStateChange);
 
   React.useEffect(() => {
-    onStateChangeRef.current = onStateChange;
     onReadyRef.current = onReady;
   });
 
@@ -344,24 +299,8 @@ function BaseNavigationContainerInner({
   }, [getRootState, state]);
 
   useClientLayoutEffect(() => {
-    const hydratedState = getRootState();
-
-    // TODO(@ubax): invesitagte if there is cleaner way to do it
-    // If not consider deprecating the prop
-    const onStateChange = onStateChangeRef.current;
-    const shouldNotifyStateChange =
-      hasNotifiedInitialStateRef.current &&
-      lastNotifiedStateRef.current !== hydratedState &&
-      onStateChange !== undefined;
-    hasNotifiedInitialStateRef.current = true;
-    lastNotifiedStateRef.current = hydratedState;
-
     emitter.emit({ type: 'state', data: { state } });
-
-    if (shouldNotifyStateChange) {
-      onStateChange(hydratedState);
-    }
-  }, [getRootState, emitter, state]);
+  }, [emitter, state]);
 
   return (
     <NavigationContainerRefContext.Provider value={navigation}>